Reworked library reservations' view
