+++ uni/lib/view/library/widgets/library_reservations_tab.dart
@@ -3,7 +3,7 @@ import 'package:flutter_redux/flutter_redux.dart';
 import 'package:tuple/tuple.dart';
 import 'package:uni/model/app_state.dart';
 import 'package:uni/model/entities/library_reservation.dart';
-import 'package:uni/view/library/widgets/library_reservations_card.dart';
+import 'package:uni/view/library/widgets/reservation_row.dart';
 
 class LibraryReservationsTab extends StatelessWidget {
   const LibraryReservationsTab({super.key});
@@ -45,7 +45,24 @@ class LibraryReservationsList extends StatelessWidget {
         scrollDirection: Axis.vertical,
         shrinkWrap: true,
         children: [
-          LibraryReservationsCard(),
+          getReservationList(context),
         ]);
+  }
+
+  Widget getReservationList(BuildContext context) {
+    final List<Widget> rooms = [];
+
+    for (int i = 0; i < reservations!.length && i < 2; i++) {
+      rooms.add(Container(
+          padding: const EdgeInsets.all(8),
+          decoration: BoxDecoration(
+              border: Border(
+                  bottom: BorderSide(
+                      color: Theme.of(context).dividerColor, width: 1))),
+          margin: const EdgeInsets.all(8),
+          child: ReservationRow(reservations![i])));
+    }
+
+    return Column(children: rooms);
   }
 }
